Updated mysql cartridge agent scripts to wait for mysql port 3306 and phpadmin 80, added a separate log for git.sh and fixed .netrc logic

+++ products/cartridge-agent/modules/cartridge-agent/ec2/php/cartridge-agent.sh
@@ -103,69 +103,111 @@ echo "Generating git.sh..." | tee -a $LOG
 # If repo is available do a git pull, else clone
 echo "#!/bin/bash
 set -e
+GIT_SH_LOG=/var/log/apache-stratos/git-sh.log
 if [ -d \"${APP_PATH}/.git\" ]; then
    cd ${APP_PATH}
-    
+   echo \"Invoking repo info service...\" | tee -a \$GIT_SH_LOG
    curl -X POST -H \"Content-Type: text/xml\" -H \"SOAPAction: urn:getRepositoryCredentials\" -d @${instance_path}/repoinforequest.xml --silent  \"${REPO_INFO_EPR}\" --insecure > /tmp/git.xml
+   echo \"Processing repo info service response...\" | tee -a \$GIT_SH_LOG
    sed '1,5d' /tmp/git.xml > /tmp/git1.xml
    sed '2d' /tmp/git1.xml > /tmp/git.xml
    username=\`xml_grep 'ax29:userName' /tmp/git.xml --text_only\`
    password=\`xml_grep 'ax29:password' /tmp/git.xml --text_only\`
    repo=\`xml_grep 'ax29:url' /tmp/git.xml --text_only\`
+   echo \"username=\$username repo=\${repo}\" | tee -a \$GIT_SH_LOG
    rm /tmp/git1.xml
    rm /tmp/git.xml
+   echo \"Preparing .netrc...\" | tee -a \$GIT_SH_LOG
    url=\`echo \$repo |sed 's/http.*\/\///g' |sed 's/\:.*//g' |sed 's/\/.*//g'\`
    echo \"machine \${url} login \${username} password \${password}\" > ~/.netrc
    sudo echo \"machine \${url} login \${username} password \${password}\" > /root/.netrc
    chmod 600 ~/.netrc
    sudo chmod 600 /root/.netrc
+   echo \"Setting git http.sslVerify false\" | tee -a \$GIT_SH_LOG 
    git config --global --bool --add http.sslVerify false
+   echo \"Running git pull...\" | tee -a \$GIT_SH_LOG
    sudo git pull  
-   rm ~/.netrc
-   sudo rm /root/.netrc
+   if [ -f ~/.netrc ]; then
+      echo \"Removing ~/.netrc...\" | tee -a \$GIT_SH_LOG
+      rm ~/.netrc
+   fi
+   if [ -f /root/.netrc ]; then
+      echo \"Removing /root/.netrc...\" | tee -a \$GIT_SH_LOG
+      sudo rm /root/.netrc
+   fi
 
+   echo \"Changing owner of application path/www to www-data:www-data...\" | tee -a \$GIT_SH_LOG
    sudo chown -R www-data:www-data ${APP_PATH}/www
-   if [ -f \"${APP_PATH}/sql/alter.sql\" ]; then
+   if [ -f \"${APP_PATH}/sql/alter.sql\" ] && [ -n \"${MYSQL_HOST}\" ] && [ -n \"${MYSQL_USER}\" ] && [ -n \"${MYSQL_PASSWORD}\" ]; then
  	mysql -h ${MYSQL_HOST} -u ${MYSQL_USER} -p${MYSQL_PASSWORD} < ${APP_PATH}/sql/alter.sql
    fi
 else
+   echo \"Removing index.html from application path...\" | tee -a \$GIT_SH_LOG
    sudo rm -f ${APP_PATH}/index.html
+   echo \"Invoking repo info service...\" | tee -a \$GIT_SH_LOG
    curl -X POST -H \"Content-Type: text/xml\" -H \"SOAPAction: urn:getRepositoryCredentials\" -d @${instance_path}/repoinforequest.xml --silent  \"${REPO_INFO_EPR}\" --insecure > /tmp/git.xml
+   echo \"Processing repo info service response...\" | tee -a \$GIT_SH_LOG
    sed '1,5d' /tmp/git.xml > /tmp/git1.xml
    sed '2d' /tmp/git1.xml > /tmp/git.xml
    username=\`xml_grep 'ax29:userName' /tmp/git.xml --text_only\`
    password=\`xml_grep 'ax29:password' /tmp/git.xml --text_only\`
    repo=\`xml_grep 'ax29:url' /tmp/git.xml --text_only\`
+   echo \"username=\$username repo=\${repo}\" | tee -a \$GIT_SH_LOG
    rm /tmp/git1.xml
    rm /tmp/git.xml
+   echo \"Preparing .netrc...\" | tee -a \$GIT_SH_LOG
    url=\`echo \$repo |sed 's/http.*\/\///g' |sed 's/\:.*//g' |sed 's/\/.*//g'\`
    echo \"machine \${url} login \${username} password \${password}\" > ~/.netrc
    sudo echo \"machine \${url} login \${username} password \${password}\" > /root/.netrc
    chmod 600 ~/.netrc
    sudo chmod 600 /root/.netrc
+   echo \"Setting git http.sslVerify false\" | tee -a \$GIT_SH_LOG
    git config --global --bool --add http.sslVerify false
+   echo \"Creating temporary git folder...\" | tee -a \$GIT_SH_LOG
    sudo mkdir ${instance_path}/temp_git
-   sudo git clone \${repo} ${instance_path}/temp_git
+   echo \"Running git clone...\" | tee -a \$GIT_SH_LOG
+   git clone \${repo} ${instance_path}/temp_git
+   echo \"Moving content to /var/www\" | tee -a \$GIT_SH_LOG
    sudo mv ${instance_path}/temp_git/* /var/www
    sudo mv ${instance_path}/temp_git/.git /var/www
    sudo rm -rf ${instance_path}/temp_git
-   rm ~/.netrc
-   sudo rm /root/.netrc
- 
-
-   if [ -f \"${APP_PATH}/sql/init.sql\" ]; then
-       mysql -h ${MYSQL_HOST} -u ${MYSQL_USER} -p${MYSQL_PASSWORD} < ${APP_PATH}/sql/init.sql
+   if [ -f ~/.netrc ]; then
+      echo \"Removing ~/.netrc...\" | tee -a \$GIT_SH_LOG
+      rm ~/.netrc
    fi
-   echo \"SetEnv STRATOS_MYSQL_USER ${MYSQL_USER}
-   SetEnv STRATOS_MYSQL_HOST ${MYSQL_HOST}
-   SetEnv STRATOS_MYSQL_PASSWORD ${MYSQL_PASSWORD}
-   \" > /tmp/.htaccess
-   sudo mv /tmp/.htaccess ${APP_PATH}/
+   if [ -f /root/.netrc ]; then
+      echo \"Removing /root/.netrc...\" | tee -a \$GIT_SH_LOG
+      sudo rm /root/.netrc
+   fi
+ 
+   if [ -n \"${MYSQL_HOST}\" ] && [ -n \"${MYSQL_USER}\" ] && [ -n \"${MYSQL_PASSWORD}\" ]; then
+      if [ -f \"${APP_PATH}/sql/init.sql\" ]; then
+          mysql -h ${MYSQL_HOST} -u ${MYSQL_USER} -p${MYSQL_PASSWORD} < ${APP_PATH}/sql/init.sql
+      fi
+      echo \"Generating .htaccess...\" | tee -a \$GIT_SH_LOG
+      echo \"SetEnv STRATOS_MYSQL_USER ${MYSQL_USER}
+      SetEnv STRATOS_MYSQL_HOST ${MYSQL_HOST}
+      SetEnv STRATOS_MYSQL_PASSWORD ${MYSQL_PASSWORD}
+      \" > /tmp/.htaccess
+      echo \"Copying .htaccess to application path...\" | tee -a \$GIT_SH_LOG
+      sudo mv /tmp/.htaccess ${APP_PATH}/
+   fi
+   echo \"Changing owner of application path/www to www-data:www-data...\" | tee -a \$GIT_SH_LOG 
    sudo chown -R www-data:www-data ${APP_PATH}/www
-    
-fi" > ${instance_path}/git.sh
+   
+fi
+echo \"git.sh done\" | tee -a \$LOG" > ${instance_path}/git.sh
 echo "git.sh generated" | tee -a $LOG
 chmod 755 ${instance_path}/git.sh
+
+
+# -----------------------------------------------------
+# Git clone
+# -----------------------------------------------------
+pushd ${instance_path}
+echo "Running git.sh..." | tee -a $LOG
+sudo sh ${instance_path}/git.sh
+echo "Git clone done" | tee -a $LOG
 
 
 # -----------------------------------------------------
@@ -176,10 +218,10 @@ do
 var=`nc -z localhost 80; echo $?`;
 if [ $var -eq 0 ]
    then
-       echo "port 80 is available" | tee -a $LOG
+       echo "Port 80 is available" | tee -a $LOG
        break
    else
-       echo "port 80 is not available" | tee -a $LOG
+       echo "Port 80 is not available" | tee -a $LOG
    fi
    sleep 1
 done
@@ -189,10 +231,10 @@ do
 var=`nc -z localhost 443; echo $?`;
 if [ $var -eq 0 ]
    then
-       echo "port 443 is available" | tee -a $LOG
+       echo "Port 443 is available" | tee -a $LOG
        break
    else
-       echo "port 443 is not available" | tee -a $LOG
+       echo "Port 443 is not available" | tee -a $LOG
    fi
    sleep 1
 done
@@ -216,12 +258,4 @@ echo "Executing: event-publisher.sh $MB_IP $MB_PORT $event_class_name $event_jso
 sh event-publisher.sh $MB_IP $MB_PORT $event_class_name $event_json_path
 echo "Event published" | tee -a $LOG
 popd
-
-# -----------------------------------------------------
-# Git clone
-# -----------------------------------------------------
-pushd ${instance_path}
-echo "running git clone..." | tee -a $LOG
-sudo sh ${instance_path}/git.sh
-echo "git clone done" | tee -a $LOG
